Moved BufferedReadTransceiver to its own file
